[Explore] Add GA Events to Page overview (#5384)
Added the following tracking events:
- Component View: A generalized event to track the initial and total
views of a component
- Page Overview Click: An event to track the clicks on the embedded
scrolling links
- Total Component View Time: An event to track the duration a component
was on screen.
- Total Anchor Count: An event to track the number of anchor elements
present in a component.

Added the following parameters:
- View Tracking Mode: A way to distinguish between tracking the initial
view and total views.
- Click Tracking Mode: A way to distinguish between tracking the first
click on a component and the total clicks of a component.
- Count Anchor Elements: Stores the total number of anchor elements
- Total View Time: Stores the total time a component was on screen.

Added the following values:
- Initial View: Indicates it is the first view of a component.
- Initial Click: Indicates it is the first click on a component element.
- Total Views: Indicates that repeated views of the same component are
being tracked on the same page.
- Total Clicks: Indicates that repeated clicks on the same type of
elements are being tracked on the same page.

For the View Time and Anchor Count events, the value given is the number
calculated in code converted to string. The string is treated as a
number by Google Analytics. See:
https://support.google.com/analytics/answer/14240153?sjid=12384612953157240702-NA&visit_id=638902601720285501-2700643701&rd=1