Metadata modal denominators (#5382)
## Description

This PR incorporates denominator (per capita) metadata into the new
metadata modal.

Previously, the old metadata modal was displaying the stat var
associated with the denominator, but no denominator facet information
was being passed through from the charts to the metadata modal.

The updates in this PR can be divided into two parts:
- Updating the bar, line, scatter and map tiles to include the
denominator facet in the set of facets that are ultimately passed to the
metadata modal.
- Updating the metadata modal to display the denominator stat var and
the associated facet information.

## Notes

With the incorporation of the denominator stat var, we now sort the stat
vars so that the regular (non-denominator) stat vars appear first,
followed by the denominator stat var if there is one. We then sort
alphabetically as a secondary sort.

The list of sources in the bar tile (next to the metadata modal trigger)
now includes the denom if it exists.

## Screenshot

<img width="757" height="955" alt="Screenshot 2025-08-07 at 5 33 54 PM"
src="https://github.com/user-attachments/assets/8c21f48f-f20d-465a-a921-ec0e0dc53d9e"
/>

## Tests

A test has been added to verify that the clicking the “per capita”
button adds the per source to the list of sources, and that inside the
metadata dialog the per capita stat var appears.

This test will be moved to shared_tests (with the other vis_timeline
tests) once the feature flag is dropped. In the meantime, it is directly
in /tests so that the cdc tests do not run it.

## Testing

This functionality is visible in any chart that has a “per capita”
checkbox, both in the NL search and in the vis tools.