[Percy] Snapshot staging (#5391)
Update Percy Snapshot diffing for the staging vs Prod runs. 
Successfully ran through all the way on friday with this setup:
https://percy.io/6fddbf6d/web/datacommons-environments-f8aa3ca5/builds/42120333/changed/2252926493?browser=chrome&browser_ids=63%2C69%2C70%2C71&group_snapshots_by=similar_diff&subcategories=unreviewed%2Cchanges_requested&viewLayout=side-by-side&viewMode=new&width=1280&widths=375%2C1280

The 300+ snapshots are sharded into 8 different parallel cloud build
steps within one action that all use the same cloud build ID for Percy
to group the snapshots as the same build.

This took <10 minutes to snapshot everything. percy still took 2 hours
to process the snapshots but I believe this is because we're on free
tier