Updates to the stat var hierarchy test_filter test to prevent intermittent failures.
The reason for the intermittent failures was that, after selecting "County", it was waiting for the loader to finish. However, `wait_for_loading` was checking against a load with id "screen", whereas this loader had an id of "browser-screen".

Because of this, the test continued right away. Usually, especially locally, this still resulted in a passing test, because the update had happened at this point. However, if the update took enough time, the race condition would be unfavorable and the test fails.

To solve this, we are now looking for the change itself, rather than the status of the spinner.

Ony once the text of the "Agriculture" element changes, we check the numbers and make sure the number has gone down.