Bar Charts - Inconsistent Facets (#5450)
## Description

This PR updates the facet selector so that when we are in "grouped" mode
(i.e., for bar charts), we no longer display inconsistent facets. An
inconsistent facet is one that does not appear across all stat vars
available in the chart.

This is row number 3. in the
[Bugbash](https://docs.google.com/spreadsheets/d/1VtmR1H3VOBbEL_D0OxP9z0ZSFmD3H6k_ZldraM11CjA/edit?gid=690212228#gid=690212228),
"Bar charts - Age distribution issues".

## Notes

I did not completely remove the previous functionality of keeping and
allowing suggestion of inconsistent facets, but rather placed that
functionality behind a flag set to false. This is for the easy recovery
of the original functionality, which may be useful for testing and
exploration of the data later.

## Screenshots

### Before

<img width="600" height="342" alt="Before"
src="https://github.com/user-attachments/assets/31cabf46-f242-4e8f-a50e-b3b2a9d556d0"
/>

### After

<img width="610" height="312" alt="After"
src="https://github.com/user-attachments/assets/3229aee1-2846-44eb-b086-5173b8e50dae"
/>

## Tests

The bar chart facet selector test needed to be updated because the chart
being tested now no longer has facet choices.

We are now testing another chart. Because this chart's sources do not
change when we change facet, we are opening up the metadata dialog to
verify the facet was changed (similar functionality to the per capita
test).

## Testing

More than one consistent facet, facet selector visible:

http://localhost:8080/explore#q=Jobs+in+the+United+States

Facet options now inconsistent, facet selector invisible.

http://localhost:8080/explore#q=Age+distribution+in+the+United+States

(This age distribution example shows the flag and inconsistent facets if
you set the `SHOW_INCONSISTENT_FACETS` flag to this.