Add short-term search indicators endpoint  (#5448)
Adds /api/nl/search-indicators endpoint which calls the nl_search_vars
endpoint to search the embeddings indexes for statvars and topics.

See
https://docs.google.com/document/d/1cACOCOzowcSndYdgDsd4S4-fi_p5l4d-GYjZGQJlsFo/edit?resourcekey=0-ySczpup9r4mrwpV5KKUIUQ&tab=t.0
for endpoint overview.

I also had an auto-save formatter on when I started this PR which caused
a lot of single quotes to be changed to double quote. This makes this
change look a lot larger than it actually is, apologies for the noise!