Webdriver Test Tightening (#5514)
## Description

This PR contains measures to tighten four tests, all of which exhibited
intermittent failures due to race conditions.

* `test_bar_select_different_facet` and `test_per_capita_metadata`:
These tests occasionally failed via the same mechanism. The presence of
the element was waited for, but this sometimes resolved before the
content tested for afterwards was ready. We resolved this by waiting for
the relevant element to be visible and with that text.
* `test_no_facet_choices_available` and `test_select_different_facet`:
These tests both assumed the existence of certain elements after the
chart finished loading (which was usually true). We now explicitly wait
for them.

## Notes

These were based on reports by @juliawu (thank you!). I could replicate
the first two issues. However, I was unable to replicate the last two.
Nevertheless, I did identify those possible routes to a race condition
in those latter tests and resolved them.