API Dialog - Custom DCs (#5522)
## Description

This PR improves the ability of the observation specs library (as used
by the API dialog) to discern between custom DCs and the standard
instance.

The previous logic relied solely on the existence of the apiRoot. If the
apiRoot existed and was not the standard DC API root, we treated it as a
custom DC.

This worked for web components, where an apiRoot was passed into the web
component in order to use a custom DC API root. However, this did not
discern between custom DCs when in the explore context. In the explore
context, the apiRoot is always undefined, and so we need another
mechanism to determine if we are in a custom DC.

### Explore Context

In the explore context, we are now relying on the `globalThis.isCustomDC
` attribute. There is an important caveat, which is that custom DCs are
not guaranteed to supply this, as this is generated in the Jinja
templates, and custom DCs usually, but do not always, set this variable.

However, we do know that the primary DC does set this, and will be the
only place that sets this to `0`. We use that knowledge to determine if
we are in a custom environment: if `isCustomDC` exists and is `0`, we
are not in a custom DC. Otherwise we are in a custom DC.

If we are not in a custom DC, we use the `DEFAULT_API_V2_ENDPOINT` as
before. Otherwise, we get the current origin from the URL, and append
the custom DC endpoint to it.

### Web Component Context

This behaves largely as it did before. We will always have an API root
(because if one is not supplied, it is set to the default API root). If
we are using the default API root, we consider ourselves not in a custom
DC. Any other API root and we are in a custom DC.

## Notes

As we were originally, we are assuming that a custom DC will have the
`/core/api` API available. This is the case on staging, but is not the
case locally when running `-e custom`. Thus locally, you will get a link
that doesn't exist in custom mode. However, you will get a valid link in
a true custom instance.

An additional minor note: I moved the two API root and path related
constants into the same constants file that contains the
DEFAULT_API_ENDPOINT (used by the web components).