[Cloud DEploy] Make Mixer deployment go through website cloudbuild (#5544)
While this isnt ideal separation - we would like to move all of this
outside of the website repo. So we will do it all at once

For now, we are just moving the cloudbuild deploy trigger to the website
repo, and commenting out the DataCommons Script Runner image rebuilding
because it needs to have submodules udpated. I will figure out the right
way to update it in a follow up.