[StatVar Autocomplete ] Only suggest queries for places we have data on (#5554)
For Stat Var Autocomplete, when there is no place identified in the
request, until now we defaulted to just show "on Earth". However, most
of our data is not at the Earth level, which caused frequent queries to
fail on StatVar Autocomplete redirection.

With this change, we don't delay the autocompletion, as Stat vars become
available in the same time frame, but we later add a randomly selected
place for which we know we have data.

See screencast:

https://screencast.googleplex.com/cast/NTI2MTk1MDk3MDc1NzEyMHw3Y2I0NWIzNS03ZQ

Note that if the user selects an autocomplete suggestions where it's
autocompleting the entire query, we redirect it to the explore page with
URL parameter injection using the SV and P params, like

http://localhost:8080/explore#p=country/USA&sv=Annual_Emissions_CarbonDioxide_FossilFuelOperations&q=Annual%20Amount%20of%20Emissions%3A%20Fossil%20Fuel%20Operations%2C%20Carbon%20Dioxide%20in%20United%20States

However, when the autocomplete suggestion only replaces a partial query
(ex: "Tell me about the life expectancy" --> The life expectancy SV will
only match the tail of the query), we execute the query via the NL
endpoint.

For selecting places, we only use the TopPlaces of type Country, State,
AA1 or EurostatNUTS1