Disable browser page observation redirect webdriver tests (#5595)
Skips `TestBrowser.test_observation_table_redirect` and
`TestBrowser.test_observation_chart_redirect` webdriver tests.

Context:

These tests are currently failing and blocking PRs. These tests check
that clicking the links in observation tables and charts on the browser
pages redirect the user to the browser page for a specific observation.

After some investigation, there appear to be no recent code changes that
would result in breaking this functionality. This functionality can be
flaky on autopush, but works in staging and prod. We suspect a recent BQ
update may have introduced some latency, causing these tests to fail. We
are disabling them for now to unblock PRs.