Prim Token Assertions (#5609)
## Description

This push adds explicit assertions before processing a Prism node as
either a string or an array.

The trigger for this is an error in the pipeline build that implies
there is a difference in how TypeScript is able to infer the type in
these two branches

## Note

I could see the same errors when running `npx tsc
js/components/elements/code/code_block.tsx --noEmit --skipLibCheck`.
This ignores our `tsconfig.json` (and thus causes a number of other
errors). However, key to effecting the patch, we can use this to verify
that type assertions on the two branches to solve the problem, at least
in that local type-check.