Facet Terminology (#5625)
## Description

This PR runs an extraction and compilation of the i18n translation so
that the compiled translation catalog contains the correct terminology
for facets.

This is a follow-up to the following PR:

[5530](https://github.com/datacommonsorg/website/pull/5530)

Screenshots and details of the actual changes are described in the
description of that earlier PR.

The reason there was this difference between the catalog and the default
names was that at the same time as 5530 was going through review and
deployment, another PR independently compiled the catalog. This added
the messages to the compiled English translations of the datasets for
the first time (but without the "facet" terminology change from this
PR).

The extraction and compilation also updated some other minor copy
differences between the catalog and the messages.