Adding a x-surface metadata tag to website API calls (#5475)
Passes an `x-surface` metadata header when making calls to the mixer
`v2/observations` endpoint, which will be used to indicate which DC
feature (website, custom DC, MCP server, etc.) the query originates
from.

This flag defaults to "website", but can be overridden if another
surface makes a call to the website by passing in the `x-surface` header
to that call. This is stored as the `surfaceHeaderValue` and is passed
through to any calls made to Mixer. For example, DataGemma makes calls
to the `nodejs/query` endpoint, which calls `/api/detect-and-fulfill`
and fetches data for each resulting chart from Mixer.

This also includes a `web-component` header for the web components that
can be embedded in other sites. These render the tiles, so this value is
passed through each tile type to their respective mixer calls.

The main functional changes are in `server/services/datacommons.py`,
most other changes are just passing in the `surfaceHeaderValue` to
functions that reach the API call to ensure this is propagated in every
case. Including this header throughout the repository in addition to the
endpoints currently used by other services makes it more clear for
future engineers that this is something that they need to consider when
using website code in a new feature. I also added an "unknown" surface
value fallback in the `get` and `push` functions that ultimately process
calls to mixer, which ensures that we'll see "unknown" in mixer's logs
if something is going wrong in the website. This differentiates this
error from an empty surface header, which we expect to see from the
public API and thus wouldn't be able to identify as an issue in the
website.

This also adds the `surfaceHeaderValue` to `args_to_ignore` in the cache
decorator on `get` and `post` so that we don't cache based on the
surface header.