[Cleanup] Remove all references to our Screenshot testing (#5626)
Screenshot testing was built with our own soltuion in GCS, shown through
the website. It has been moved over to using Percy.

In this PR we remove all references to the website's "/screenshots"
pages, remove the configurations to run the run_test.sh -s mode which
runs screenshots.
Removes the references to the screenshotDomain in the GKE Cron Config.
Removes all other references to screenshots that we could find.