[Webdriver Tests] Cleanup Map Tests (#5643)
Per discussion with @gmechali and as followup to #5642, this PR makes a
number of adjustments to the map webdriver tests

(1) Per discussion with @gmechali, this PR splits find_element(s)
operations by `CSS_SELECTOR` into sequential finds using `ID` and
`TAG_NAME` selectors. The goal is to maintain readability, with the
understanding that the sequential calls should not result in stale
element errors.

(2) Memoizes the dataCommonsClient, which was causing a loop of chart
re-renders when the client was reinitialized on each render. This should
prevent tests from breaking when the chart re-renders mid-test.