[Webdriver Tests] Add wait to test_bar_select_different_facets (#5644)
The `TestExplorePage.test_bar_select_different_facets` webdriver test
sometimes fails with:

```
AssertionError: unexpectedly None : Could not find the 'Categories of Jobs' chart block.
```

This change makes the test less flakey by having the test explicitly
wait for the condition that the "Categories of Jobs" block is found
before moving on with the test.