[Vis Tools] Fail more elegantly on failing scatter plot fetches (#5636)
# Context
For some place and place type combinations, the scatter plot tool shows
a "still loading..." spinner that never ends, and a flurry of error
alerts that prevent further interaction with the page.

Steps to replicate:
1. Go to
[datacommons.org/tools/scatter](http://datacommons.org/tools/scatter)
1. Enter "India" into the search bar
1. Select AdministrativeArea4 for place type

The cause is a failing V2 Node API fetch for enclosed places that isn't
appropriately caught and handled.

# Changes
This PR updates the error handling to "fail more elegantly" when this
fetch fails, by making these changes:
(1) Break the infinite loop of failed fetches. Adds a new state that
saves the last failed fetch and skips any requests that would repeat
that fetch.
(2) Add an abort controller to cancel any old fetches when the component
re-renders.
(3) Only make the backup fetch if the named fetch fails, instead of
concurrently.
(4) Update client tests so that mocked axios calls accept a second
argument (the abort signal)

# Screengrabs

## Before


https://github.com/user-attachments/assets/750055bf-f237-45a1-8fe2-1e9d3f323ec0


## After


https://github.com/user-attachments/assets/eeb583da-8725-4ac2-9121-3ba7f494c413