[Vis Tools] Dynamically show and hide stat var hierarchy (#5637)
To better guide users through using our visualization tools, this PR
starts the stat var hierarchy collapsed, and only shows the hierarchy if
a place (& place type for map/scatter) is selected.

More specifically, the following changes are made:
(1) Add a new react hook that allows a state to be optionally
controllable by a parent element
(2) Add logic to /tools/map, /tools/scatter, and /tools/timeline to show
and hide the stat var hierarchy depending on whether or not a place is
selected.
(3) Adjust the CSS styling on StatVarWidget so that transitioning from
collapsed -> visible will use a smooth transition instead of popping in
instantly.

# Screen grabs


https://github.com/user-attachments/assets/7964a684-bd68-43ff-955c-adbfff604a10



https://github.com/user-attachments/assets/33091f34-d866-4bc2-a710-da205d2fae88



https://github.com/user-attachments/assets/807e7701-f022-4fc0-b242-09565a9cc45d