Facets in downloads (#5653)
## Description

This PR addresses an issue in the download tool, referenced on line 17
of the [UX
items](https://docs.google.com/spreadsheets/d/1VtmR1H3VOBbEL_D0OxP9z0ZSFmD3H6k_ZldraM11CjA/edit?gid=592668190#gid=592668190)

The issue was that with the introduction of facet selection to the chart
tiles, the Download tool's results no longer consistently matched the
chart results.

This can be demonstrated in master, using any chart where facet
selection is available. For example, by doing the following:
* Search for [Jobs in
Texas](http://localhost:8080/explore?enable_feature=enable_stat_var_autocomplete#q=Jobs+in+Texas&client=ui_query).
* The first chart will be Categories of Jobs.
* Click "Download" and download the CSV (you can also look at the data
in the window, but the CSV is easier to parse)
* The data here correctly match the data in the chart, because both the
chart and the download are using the best available data.
* Click "Explore other facets" and select the yearly data `Yearly
(P1Y)`.
* The chart will change, and you can verify that the chart is using the
yearly data by mousing over the bars.
* Click "Download" and download the CSV. You will see that the download
is still using the "best available data" (which is monthly).

The issue came about because the implementation of the data download
functionality was done before explicit facet selection was available.

### Per Capita Variables

Related to the primary goal of allowing facet selection through the Data
Commons Client library, the Data Commons client library did not allow
for custom demon variables tied to specific stat vars. In other words,
if a particular chart used something other than Count_Person (for
example, Count_Household), the charts themselves would use the custom
denominator, but the Data Commons Client always used Count_Person.

## The Solution

The Data Commons Client's function interfaces previously took a list of
DCIDs through the `variable` parameter, and use those to construct the
payload. This method of interfacing with the functions is still allowed
for flexibility and backwards-compatibility.

We now also allow variables to be sent in directly as a list stat var
specs. This gives the Data Commons Client enough information to do both
facet-specific selection and custom denom variables.

If stat var specs are provided, the functions will use those. Otherwise,
it will fall back to using the variables and behave as it did before.

With the delivery of stat var specs, we then group them by facet id, so
that we can make as few API calls as necessary, and merge the results of
those calls into a final payload.

While the updates to the Data Commons Client is fairly substantial, the
updates to each of the four chart types (line, scatter, map, ranking)
are relatively minor. The highlight chart does not have a download
button and so is not involved.

## Notes

While the introduction of facet capability into the Data Commons client
was initially envisioned as a relatively small task, it turned out to
require a fair amount of adjustments to a number of functions within the
client.

However, these updates bring us much closer to a broader goal that we
have, of having all charts be able to do their data fetching through the
abstraction provided by the Data Commons client, rather than
individually.