Tooltips - Mobile tooltip on action (#5663)
## Description

There is an inherent challenge with applying tooltips to action elements
such as buttons and links.

For desktop uses, these tooltips can operate as normal, triggering on
mouseover (with the button or link action being applied on click as per
normal).

However, mobile users do not have a "hover" action in the same way, and
must interact with the element using some variation of a tap.

The initial version of the tooltip component recommended disabling
mobile interactions for action buttons.

This new version offers an alternative that allows the user to see the
tooltip: the [long
press](https://m2.material.io/components/tooltips#placement) described
under the Material Design specification. The link is from Material V2
because it gives a better explanation, but Material 3 suggests it as
well.

With `longPress` enabled, the tooltip will now show the tooltip after
the user holds on the element for more than a certain amount of time
(this is configurable, but 500ms by default).


## Notes

As with all tooltips, mission critical information should not be
displayed in the tooltip itself, under the assumption that a user may
not view the content of the tooltip. The information should always
augment the experience rather than be required.

## Testing

This new functionality is now applied to the PlaceAndTypeOptions
component, and can be used by going to the following chart:

[Map
Chart](http://localhost:8080/tools/scatter#%26svx%3DLifeExpectancy_Person%26dx%3DCount_Person%26svy%3DFertilityRate_Person_Female_AsAFractionOf_Count_Person_Female%26dy%3DCount_Person%26epd%3Dcountry%2FIND%26ept%3DAdministrativeArea1%26qd%3D1%26dd%3D1%26pp%3D)

The tooltip is applied to the two buttons that toggle between the
scatter and map rendering of the data. In desktop, the tooltip will
appear on hover as before. In mobile, a long press will open the
tooltip.

Another way to test tooltips in general is to temporarily show the
DemoTooltip component on the home page by adding the following section
to `static/js/apps/homepage/app.tsx`:

```
<Section>
  <DemoTooltips />
</Section>
```

## Screenshots

### Video Demo


https://github.com/user-attachments/assets/e4290aa6-7985-45ee-af00-30f814bdb94c