Wrapping Flask cache with a logger (#5652)
Creates a wrapper around `@cache.cached` and `@cache.memoize` that logs
any `requestIds` found in the cached result. These IDs uniquely identify
each Mixer response and are added to the `ObservationResponse` in [this
PR](https://github.com/datacommonsorg/mixer/pull/1660).

These logs are used to track usage from the website cache in my Mixer
usage logs. The logger writes the `requestIds` in a JSON object that can
be ingested by GCP cloud logging, which I plan to route to a BigQuery
table and join with my existing mixer logs to track which mixer calls
are used even when they are cached. The request IDs are written for
cache hits and misses -- in my logger, I plan to subtract one call from
each ID found to account for the first cache miss.

This first PR adds the cacher and uses it in `get` and `post_wrapper`,
where we most directly fetch the Mixer results. It will later be used
anywhere we cache an object that is populated by Mixer results. These
first cases use `cache.memoize`, but I added support for `cache.cached`
as well because it is used by these later sites. The later changes are
made in [this PR](https://github.com/datacommonsorg/website/pull/5693).