Add comment to clarify that the field used to recognize logs should not be changed (#5734)
The `website_cache_usage_logs` [log
router](https://pantheon.corp.google.com/logs/router?e=13803378&mods=-monitoring_api_staging&project=datcom-website-prod)
that ingests the logs in the website cache looks specifically for the
"mixer_respones_ids" in the logged response. This adds a comment to the
places where we define this header to make is more clear that this
should not be changed without also updating the log router.

This comment is also added to where we log the Mixer usage logs in [this
PR](https://github.com/datacommonsorg/mixer/pull/1685).