Highlight chart: Facet Selection (#5749)
## Description

The highlight chart had become disconnected from the facet selector,
because facetSelector, as an object, was always "true", resulting in the
[statVarSpec.facetId] no longer being applied (the original
`highlightFacet` would be undefined, and so evaluate to false).

This is an initial update to restore the connection with the facet
selector.

## Testing

[Population of the United
States](http://localhost:8080/explore#q=population+of+the+united+states&client=ui_query)

The first block of the above query is a timeline/highlight block. Before
this PR, selecting a facet will not change the highlight tile. With this
PR, the facet selection will change the highlight tile.

## Notes

I wanted to get this in quickly to restore the ability to select a facet
and have that apply to the highlight chart. I did so in a way that kept
the functionality similar to how it was before, in terms of priority
between query string selected options and facet selector selected
options.

With the old `highlightFacet`, if we had a highlightFacet, that would
take precedence. If we did not, the `statVarSpec.facetId` would take
precedence.

With this update, the logic is largely the same. We look inside
`facetSelector` and, if it has anything inside `facetMetadata`, we do
not apply the user-selected facet Id (if `facetSelector` has a year but
no `facetMetadata`, we still apply the user-seleced facet, as those two
don't conflict in the same way).

This could be a point of discussion for a subsequent PR: if a user has
selected a facet, and some facet criteria is given on the query string
(and transmitted into the tile via `facetSelector`, which should take
precedence?

A case could be made for the user-selected facet always taking
precedence. I didn't do that here because that wasn't what the logic was
before, but we could easy do that by changing the ternary.