[FIXIT] Cleanup Production Feature Flags (#5772)
Deleting feature flags in production that are no longer needed.
This is step 1 to delete flags in other environments (we cannot have
flags in production that dont exist elsewhere)

- enable_gemini_2_5_flash - code referencing this has been deleted 2
months ago
- follow_up_questions_experiment is enabled on all the same environments
as follow_up_questions_ga which shortcircuits the conditions.
- page_overview_experiment, same but with page_overview_ga
- page_overview_links is never enabled, and is for a feature hidden
behind page_overview_ga which is never enabled as well. The code for
removing it is in a separate PR.
- show_api_modal has been removed from the codebase months ago
- vai_medium_relevance_threshold is disabled everywhere - code for it is
removed in another PR.

Grouping all these production flag changes in one so that we only
restart the server ONCE.
All the separate PRs will affect non-prod environments