[FIXIT] Cleanup VAI feature flags (#5768)
The VAI Medium Threshold feature flag was used for experimenting but it
turns out to be too high - we will not go for it. Deprecating it now,
since it was never enabled in any environment, it is safe to remove the
feature flags at the same time as we rip out the code.
Production will be done separately.