[FIXIT] Clean up Gemini NL Fallback Feature Flags (#5770)
The enable_gemini_2_5_flash_lite flag was created but never enabled. We
no longer need it.
The enable_gemini_2_5_flash flag is now enabled in all environments but
references to it in the code were deleted in
https://github.com/datacommonsorg/website/commit/7102bd9d2cb327bcd26dac3b0918ae5d5cbbbd81
(2 months ago).

It is safe to remove all those feature flags, Gemini 2.5 Flash will
remain the fallback model.
Removal of enable_gemini_2_5_flash in production will be done in a
separate PR