[Vis Tools] Make "log scale" option clickable (#5867)
This PR makes it so that clicking on the label "log scale" in the
scatter tool now toggles the check box. This matches the behavior of the
"Per Capita" label.

This was fixed by making the "Input" component a child of the "Label"
component, instead of a sibling, matching the structure of the "Per
Capita" label.

<img width="1706" height="1328" alt="image"
src="https://github.com/user-attachments/assets/2093b2ee-ea7c-4a1c-8f1a-e19a4359e98f"
/>